include dsc in message
Sigh, dsc messages may not include their own checksums (for obvious
tail-eating reasons) but said checksums are needed for including in
apt source repos, so calculate them on the fly.

Also rename the `files` property to `source_files` so that it does
not collide with the Files: header in the dsc message.

Update README appropriately